requirements.txt
openai-whisper
moviepy
torch
ffmpeg-python
